+++ next-docs/pages/core/carousel.tsx
@@ -3,9 +3,11 @@ import Preview from '../../components/codePreview/Preview';
 import ComponentAnatomy from '../../components/ComponentAnatomy';
 import ComponentPageDescription from '../../components/ComponentPageDescription';
 import PropsTable from '../../components/PropsTable';
+import Autoslide from '../../public/examples/carousel/Autoslide';
 import CustomizedArrow from '../../public/examples/carousel/CustomizedArrow';
 import Default from '../../public/examples/carousel/Default';
 import Indicators from '../../public/examples/carousel/Indicators';
+import SelectIndex from '../../public/examples/carousel/SelectIndex';
 import Spaces from '../../public/examples/carousel/Spaces';
 import VisibleIndex from '../../public/examples/carousel/VisibleIndex';
 import useExamples from '../../utils/useExamples';
@@ -60,6 +62,16 @@ export default function PageCarousel() {
         title="Visible indices"
         preview={<VisibleIndex />}
         code={examples ? examples.VisibleIndex : 'Loading'}
+      />
+      <Preview
+        title="Select index"
+        preview={<SelectIndex />}
+        code={examples ? examples.SelectIndex : 'Loading'}
+      />
+      <Preview
+        title="Autoslide"
+        preview={<Autoslide />}
+        code={examples ? examples.Autoslide : 'Loading'}
       />
       <PropsTable
         title="Carousel props"
@@ -133,6 +145,13 @@ export default function PageCarousel() {
             required: false,
             default: '-',
             description: 'Index of the last visible item',
+          },
+          {
+            name: 'autoSlideDelay',
+            type: 'number',
+            required: false,
+            default: '-',
+            description: 'Interval of auto sliding in milliseconds. No auto sliding if undefined',
           },
         ]}
       />