+++ packages/select/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @heathmont/moon-select
 
+## 9.20.1
+
+### Patch Changes
+
+- Bump fixes
+- Updated dependencies [undefined]
+  - @heathmont/moon-core@9.20.1
+  - @heathmont/moon-themes@9.20.1
+  - @heathmont/moon-utils@9.20.1
+
 ## 9.20.0
 
 ### Minor Changes