+++ workspaces/templates/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @heathmont/moon-templates
 
+## 9.20.1
+
+### Patch Changes
+
+- Bump fixes
+
 ## 9.20.0
 
 ### Minor Changes